[FLINK-20098] Don't add flink-connector-files to flink-dist
Before, we had both flink-connector-files and flink-connector-base as
dependencies of flink-dist.

This implies, that users should use the dependency like this:

<dependency>
	<groupId>org.apache.flink</groupId>
	<artifactId>flink-connector-files</artifactId>
	<version>${project.version}</version>
	<scope>provided</scope>
</dependency>

which differs from other connectors where users don't need to specify
<scope>provided</scope>.

Also, flink-connector-files has flink-connector-base as a provided
dependency, which means that examples that use this dependency will not
run out-of-box in IntelliJ because transitive provided dependencies will
not be considered.

This change removes the dependencies from flink-dist and lets users use
the File Connector like any other connector.

I believe the initial motivation for "providing" the File Connector in
flink-dist was to allow us to use the File Connector under the hood in
methods such as StreamExecutionEnvironment.readFile(...). We could
decide to deprecate and remove those methods or re-add the File
Connector as an explicit (non-provided) dependency again in the future.